--- pages/pages/other-pages/about-us-1.js
@@ -1,21 +1,0 @@
-import React from "react";
-
-import NavbarThree from "../../../layout/headers/NavbarOne";
-
-
-import FooterThree from "../../../layout/footers/FooterThree";
-import Breadcrumb from "../../../layout/Breadcrumb/Breadcrumb";
-import BodyContent from "../../../components/pages/otherPages/aboutUs1";
-
-const AboutUs1 = () => {
-  return (
-    <>
-      <NavbarThree />
-      <Breadcrumb />
-      <BodyContent />
-      <FooterThree />
-    </>
-  );
-};
-
-export default AboutUs1;
